feat(bcgov/storybook): update hero image
update to mdx format

#294
