+++ packages/bcgov-theme/stories/heroimage/args.ts
@@ -1,0 +1,5 @@
+export const argTypes = {
+  url: {
+    description: 'The url of the image to use.',
+  },
+};
